Adding support for pytorch 1.7 (#409)
* Adding support for pytorch 1.7

* Updating the tests for py 1.7

* Addressed the review comments to use the functions in utils.py
