+++ tests/profiler/pytorch/test_pytorch_profiler_rnn.py
@@ -7,12 +7,12 @@ import time
 import pytest
 import torch
 import torch.nn as nn
-from tests.profiler.pytorch.utils import is_pt_1_5, is_pt_1_6
 
 # First Party
 import smdebug.pytorch as smd
 from smdebug.profiler import LocalAlgorithmMetricsReader
 from smdebug.profiler.profiler_config_parser import ProfilerConfigParser
+from smdebug.pytorch.utils import is_pt_1_5, is_pt_1_6, is_pt_1_7
 
 
 class RNN(nn.Module):
@@ -75,6 +75,6 @@ def test_pytorch_profiler_rnn(pytorch_profiler_config_parser, out_dir):
     print(f"Number of events {len(events)}")
     if is_pt_1_5():
         assert len(events) <= 64
-    elif is_pt_1_6():
+    elif is_pt_1_6() or is_pt_1_7():
         assert len(events) <= 85
     shutil.rmtree(out_dir, ignore_errors=True)
